Proper qp tag type

+++ simulation/src/applications/helper/rdma-client-helper.cc
@@ -30,7 +30,7 @@ RdmaClientHelper::RdmaClientHelper ()
 }
 
 RdmaClientHelper::RdmaClientHelper (uint16_t pg, Ipv4Address sip, Ipv4Address dip, uint16_t sport, uint16_t dport, uint64_t size, uint32_t win, uint64_t baseRtt,
-        void (*msg_handler)(void* fun_arg), void* fun_arg, int tag, int src, int dest)
+        void (*msg_handler)(void* fun_arg), void* fun_arg, uint64_t tag, int src, int dest)
 {
 	m_factory.SetTypeId (RdmaClient::GetTypeId ());
 	SetAttribute ("PriorityGroup", UintegerValue (pg));
